+++ fiboa_cli/convert.py
@@ -1,9 +1,10 @@
 import importlib
 import os
+from .util import log
 
 IGNORED_DATASET_FILES = ["__init__.py", "template.py"]
 
-def convert(dataset, output_file, cache_file = None, source_coop_url = None, collection = False, compression = None):
+def convert(dataset, output_file, cache = None, source_coop_url = None, collection = False, compression = None):
     if dataset in IGNORED_DATASET_FILES:
         raise Exception(f"'{dataset}' is not a converter")
     try:
@@ -11,7 +12,14 @@ def convert(dataset, output_file, cache_file = None, source_coop_url = None, col
     except ImportError as e:
         raise Exception(f"Converter for '{dataset}' not available or faulty: {e}")
 
-    converter.convert(output_file, cache_file = cache_file, source_coop_url = source_coop_url, collection = collection, compression = compression)
+    if (not hasattr(converter, "SOURCES") or converter.SOURCES is None) and not cache:
+        log("Data access is restricted. You need to manually get the data from the source.", "warning")
+        if hasattr(converter, "DATA_ACCESS"):
+            log("Instructions for data access:", "warning")
+            log(converter.DATA_ACCESS, "info")
+        raise Exception("Provide the folder that contains the downloaded files through the `-c` parameter.")
+
+    converter.convert(output_file, cache = cache, source_coop_url = source_coop_url, collection = collection, compression = compression)
 
 def list_all_converter_ids():
     datasets = importlib.import_module(".datasets", package="fiboa_cli")