+++ fiboa_cli/datasets/be_vlg.py
@@ -1,13 +1,8 @@
-import os
-
-from tempfile import TemporaryDirectory
-from zipfile import ZipFile
-
-from ..util import download_file, log
 from ..convert_utils import convert as convert_
 
-URI = "https://www.landbouwvlaanderen.be/bestanden/gis/Landbouwgebruikspercelen_2023_-_Definitief_(extractie_28-03-2024)_GPKG.zip"
-FILENAME = "Landbouwgebruikspercelen_2023_-_Definitief_(extractie_28-03-2024).gpkg"
+SOURCES = {
+  "https://www.landbouwvlaanderen.be/bestanden/gis/Landbouwgebruikspercelen_2023_-_Definitief_(extractie_28-03-2024)_GPKG.zip": ["Landbouwgebruikspercelen_2023_-_Definitief_(extractie_28-03-2024).gpkg"]
+}
 
 ID = "be_vlg"
 TITLE = "Crop field boundaries for Belgium - Flanders"
@@ -76,37 +71,27 @@ MISSING_SCHEMAS = {
 
 
 # Conversion function, usually no changes required
-def convert(output_file, cache_file = None, source_coop_url = None, collection = False, compression = None):
-    # We need to extract manually because the GeoPackage is zipped and in a folder, not at the root
-    log("Loading file from: " + URI)
-    path = download_file(URI, cache_file)
-
-    with TemporaryDirectory() as tmp_dir:
-        log("Unzipping to: " + tmp_dir)
-        with ZipFile(path, 'r') as f:
-            f.extractall(tmp_dir)
-
-            new_path = os.path.join(tmp_dir, FILENAME)
-            convert_(
-                output_file,
-                cache_file,
-                new_path,
-                COLUMNS,
-                ID,
-                TITLE,
-                DESCRIPTION,
-                BBOX,
-                provider_name=PROVIDER_NAME,
-                provider_url=PROVIDER_URL,
-                source_coop_url=source_coop_url,
-                extensions=EXTENSIONS,
-                missing_schemas=MISSING_SCHEMAS,
-                column_additions=ADD_COLUMNS,
-                column_migrations=COLUMN_MIGRATIONS,
-                column_filters=COLUMN_FILTERS,
-                migration=MIGRATION,
-                attribution=ATTRIBUTION,
-                store_collection=collection,
-                license=LICENSE,
-                compression=compression
-            )
+def convert(output_file, cache = None, source_coop_url = None, collection = False, compression = None):
+    convert_(
+        output_file,
+        cache,
+        SOURCES,
+        COLUMNS,
+        ID,
+        TITLE,
+        DESCRIPTION,
+        BBOX,
+        provider_name=PROVIDER_NAME,
+        provider_url=PROVIDER_URL,
+        source_coop_url=source_coop_url,
+        extensions=EXTENSIONS,
+        missing_schemas=MISSING_SCHEMAS,
+        column_additions=ADD_COLUMNS,
+        column_migrations=COLUMN_MIGRATIONS,
+        column_filters=COLUMN_FILTERS,
+        migration=MIGRATION,
+        attribution=ATTRIBUTION,
+        store_collection=collection,
+        license=LICENSE,
+        compression=compression
+    )