+++ fiboa_cli/datasets/de_nds.py
@@ -1,6 +1,6 @@
 from ..convert_utils import convert as convert_
 
-URI = "https://sla.niedersachsen.de/mapbender_sla/download/FB_NDS.zip"
+SOURCES = "https://sla.niedersachsen.de/mapbender_sla/download/FB_NDS.zip"
 ID = "de_nds"
 TITLE = "Field boundaries for Lower Saxony, Germany"
 DESCRIPTION = """A field block (German: "Feldblock") is a contiguous agricultural area surrounded by permanent boundaries, which is cultivated by one or more farmers with one or more crops, is fully or partially set aside or is fully or partially taken out of production."""
@@ -38,13 +38,13 @@ MISSING_SCHEMAS = {
     }
 }
 
-def convert(output_file, cache_file = None, source_coop_url = None, collection = False, compression = None):
+def convert(output_file, cache = None, source_coop_url = None, collection = False, compression = None):
     """
     Converts the Lower Saxony (Germany) field boundary datasets to fiboa.
     """
     convert_(
-        output_file, cache_file,
-        URI, COLUMNS, ID, TITLE, DESCRIPTION, BBOX,
+        output_file, cache,
+        SOURCES, COLUMNS, ID, TITLE, DESCRIPTION, BBOX,
         license=LICENSE,
         extensions=EXTENSIONS,
         missing_schemas=MISSING_SCHEMAS,