+++ tests/test_convert.py
@@ -1,15 +1,12 @@
 import re
-from pytest import fixture, mark
-from pathlib import Path
-from fiboa_cli.util import load_file
-from fiboa_cli import convert, validate
 import tempfile
+from pytest import fixture, mark
+from fiboa_cli import convert, validate
 from click.testing import CliRunner
 
 """
 Create input files with: `ogr2ogr output.gpkg -limit 100 input.gpkg`
 """
-
 
 @fixture
 def out_file():
@@ -17,11 +14,9 @@ def out_file():
         yield out
 
 
-@mark.parametrize("converter", ["nl", "nl_crop", "be_vlg", "fr"])
+@mark.parametrize("converter", ["at", "be_vlg", "de_sh", "fr", "nl", "nl_crop"])
 def test_converter(out_file, converter):
-    path = next(Path("tests/data-files").glob(f"{converter}.*")).as_posix()
-    assert load_file(path), f"Input file missing: {path}"
-
+    path = f"tests/data-files/convert/{converter}"
     runner = CliRunner()
     result = runner.invoke(convert, [converter, '-o', out_file.name, '-c', path])
     assert result.exit_code == 0, result.output
@@ -31,3 +26,4 @@ def test_converter(out_file, converter):
 
     result = runner.invoke(validate, [out_file.name, '--data'])
     assert result.exit_code == 0, result.output
+